# 8265411: Avoid unnecessary Method::init_intrinsic_id calls