--- a/src/hotspot/share/oops/method.cpp
+++ b/src/hotspot/share/oops/method.cpp
@@ -1404,7 +1404,7 @@ methodHandle Method::make_method_handle_intrinsic(vmIntrinsics::ID iid,
   assert(MethodHandles::is_signature_polymorphic_name(m->name()), "");
   assert(m->signature() == signature, "");
   m->compute_from_signature(signature);
-  m->init_intrinsic_id();
+  m->init_intrinsic_id(klass_id_for_intrinsics(m->method_holder()));
   assert(m->is_method_handle_intrinsic(), "");
 #ifdef ASSERT
   if (!MethodHandles::is_signature_polymorphic(m->intrinsic_id()))  m->print();
@@ -1566,14 +1566,14 @@ vmSymbolID Method::klass_id_for_intrinsics(const Klass* holder) {
   }
 }
 
-void Method::init_intrinsic_id() {
+void Method::init_intrinsic_id(vmSymbolID klass_id) {
   assert(_intrinsic_id == static_cast<int>(vmIntrinsics::_none), "do this just once");
   const uintptr_t max_id_uint = right_n_bits((int)(sizeof(_intrinsic_id) * BitsPerByte));
   assert((uintptr_t)vmIntrinsics::ID_LIMIT <= max_id_uint, "else fix size");
   assert(intrinsic_id_size_in_bytes() == sizeof(_intrinsic_id), "");
 
   // the klass name is well-known:
-  vmSymbolID klass_id = klass_id_for_intrinsics(method_holder());
+  assert(klass_id == klass_id_for_intrinsics(method_holder()), "must be");
   assert(klass_id != vmSymbolID::NO_SID, "caller responsibility");
 
   // ditto for method and signature:
@@ -1593,11 +1593,6 @@ void Method::init_intrinsic_id() {
 
   vmIntrinsics::ID id = vmIntrinsics::find_id(klass_id, name_id, sig_id, flags);
   if (id != vmIntrinsics::_none) {
-#if 0
-    ResourceMark rm;
-    char buff[200];
-    log_info(cds,symboltable)("intrinsics %s = %s", external_name(), vmIntrinsics::short_name_as_C_string(id,  buff, sizeof(buff)));
-#endif
     set_intrinsic_id(id);
     if (id == vmIntrinsics::_Class_cast) {
       // Even if the intrinsic is rejected, we want to inline this simple method.