`check_methods_for_intrinsics()` in classFileParser.cpp calls `Method::klass_id_for_intrinsics()` to see if a class has intrinsic methods. However, the latter returns any class whose name is included in `vmSymbols`. This causes many unnecessary calls to `Method::init_intrinsic_id()`.

To fix this, we precompute the classes that are known to have intrinsics using `constexpr`. See `vmIntrinsics::class_has_intrinsics`.

After the fix, when running `java -Xshare:off -version`, the number of classes initialized for intrinsics is reduced from 130 to 30.

Testing: mach5 tiers 1~4 passed.

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [x] Change must be properly reviewed

### Issue
 * [JDK-8265411](https://bugs.openjdk.java.net/browse/JDK-8265411): Avoid unnecessary Method::init_intrinsic_id calls


### Reviewers
 * [David Holmes](https://openjdk.java.net/census#dholmes) (@dholmes-ora - **Reviewer**) ⚠️ Review applies to eb8e0bf693cdce260ac17ae83136aae00da1257e
 * [Yumin Qi](https://openjdk.java.net/census#minqi) (@yminqi - **Reviewer**) ⚠️ Review applies to eb8e0bf693cdce260ac17ae83136aae00da1257e


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/3564/head:pull/3564` \
`$ git checkout pull/3564`

Update a local copy of the PR: \
`$ git checkout pull/3564` \
`$ git pull https://git.openjdk.java.net/jdk pull/3564/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 3564`

View PR using the GUI difftool: \
`$ git pr show -t 3564`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/3564.diff">https://git.openjdk.java.net/jdk/pull/3564.diff</a>

</details>