@iklam 2021-04-20
> Hi Ioi,
> 
> That all appears fine to me.
> 
> Is there any startup impact?
> 
> Thanks,
> David

Hi David, when CDS is turned off, I can see a small improvement in start-up, due to the reduction in intrinsic ID lookup:

```
$ perf stat -r 100 bin/java -Xshare:off -Xint -version

OLD: 191,632,128 cycles 0.074624 sec
NEW: 191,387,603 cycles 0.073704 sec
```

With CDS enabled, the intrinsic ID lookup is done during archive dump time, so this PR makes no difference.

Also, the lookup table itself is generated during C++ compilation, so it has no runtime impact:

```
	.section	.rodata
	.type	_ZL18_intrinsics_lookup, @object
_ZL18_intrinsics_lookup:
	.byte	0
	.byte	0
	.byte	1
	.byte	1
        ....
```