@yminqi 2021-04-20
BTW, the the fix works for the problem: adding symbols to vmSymbols.hpp caused debug build failed due this bug.